Improve KEY word
When pasting files into the terminal, KEY and ACCEPT do not
handle DOS or UNIX line-endings very well. It expects a CR and
only a CR (A.K.A MAC line-ending).

So, If we receive a LF, after previously receiving a CR, we
count that as a DOS line-ending and then loop back to fetch
the next character.

If we receive a LF, without previously receiving a CR, then
we count that as a Unix line-ending, and act as if it was
a CR.